feat: create FifthInputTicketView
